requirements.txt
tensorflow>=2.16.1
keras>=3.0.0
streamlit>=1.32.0
pillow==10.2.0
numpy==1.26.3
matplotlib==3.8.2
gdown==5.1.0